Refactor render package
