+++ internal/operator-controller/rukpak/render/registryv1/generators/resources.go
